# docs/api/openapi.json
{
  "components": {
...
    "schemas": {
...
      "AuthenticatedSession": {
        "properties": {
          "assurance": {
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "sessionToken": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "AuthenticationRequest": {
        "properties": {
          "loginIdentifier": {
            "maxLength": 64,
            "minLength": 3,
            "pattern": "[A-Za-z0-9._-]+",
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "loginIdentifier",
          "password"
        ],
        "type": "object"
      },
...
      "IssuedKeyView": {
        "properties": {
          "alreadyIssued": {
            "type": "boolean"
          },
          "keyId": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "KeyView": {
        "properties": {
          "issuedAt": {
            "type": "string"
          },
          "keyId": {
            "type": "string"
          },
          "revokedAt": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "MerchantStandingRequest": {
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
      "MerchantView": {
        "properties": {
          "displayName": {
            "type": "string"
          },
          "merchantId": {
            "type": "string"
          },
          "settlementCurrency": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "MfaConfirmationRequest": {
        "properties": {
          "code": {
            "maxLength": 8,
            "minLength": 6,
            "pattern": "[0-9]+",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The public HTTP contract. Versioned in the path (ADR-0015); errors follow RFC 9457, catalogued in docs/architecture/ERROR_CONTRACT.md. Generated from the application on every build and compared against this committed copy, so a change to the contract cannot reach a client without being reviewed.",
    "title": "finapp platform API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/authentications": {
      "post": {
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedSession"
                }
              }
            },
            "description": "Created"
          }
        }
      }
    },
...
    "/v1/me/mfa/confirmation": {
      "post": {
        "operationId": "confirmMfaEnrolment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/v1/operator/merchants/{id}": {
      "get": {
        "operationId": "viewMerchant",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantView"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/v1/operator/merchants/{id}/api-keys": {
      "get": {
        "operationId": "listApiKeys",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KeyView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        }
      },
      "post": {
        "operationId": "issue",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedKeyView"
                }
              }
            },
            "description": "Created"
          }
        }
      }
    },
    "/v1/operator/merchants/{id}/api-keys/{keyId}": {
      "delete": {
        "operationId": "revoke",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "keyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantStandingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  }
}